feat: changed all the loading screens indication to button loading indication and enable user to edit product in the create receipt page
