+++ digital_receipt/lib/main.dart
@@ -52,7 +52,8 @@ Future<dynamic> myBackgroundMessageHandler(Map<String, dynamic> message) async {
 
 void main() => runApp(DevicePreview(
       builder: (_) => MyApp(),
-      enabled: !kReleaseMode,
+      // TODO replace with !kReleaseMode
+      enabled: false,
     ));
 
 class MyApp extends StatelessWidget {
@@ -132,13 +133,11 @@ class _ScreenControllerState extends State<ScreenController> {
     print('id: $id');
   }
 
- 
-
   @override
   void initState() {
     super.initState();
     initConnect();
-    
+
     initSharedPreferenceDb();
     getCurrentAutoLogoutStatus();
 
@@ -200,7 +199,6 @@ class _ScreenControllerState extends State<ScreenController> {
         //INSERTING NOTIFICATION TO SQFLITE DB
       },
       onResume: (Map<String, dynamic> message) async {
-        
         print("onResume: $message");
 
         //INSERTING NOTIFICATION TO SQFLITE DB
@@ -247,5 +245,3 @@ class _ScreenControllerState extends State<ScreenController> {
         });
   }
 }
-
-
