+++ digital_receipt/lib/screens/forgot_password.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
 import 'package:digital_receipt/services/api_service.dart';
+import 'package:digital_receipt/widgets/button_loading_indicator.dart';
 import 'package:digital_receipt/widgets/loading.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
@@ -17,131 +18,135 @@ class ForgotPassword extends StatefulWidget {
 class _ForgotPasswordState extends State<ForgotPassword> {
   GlobalKey<FormState> _formKey = GlobalKey<FormState>();
 
-  bool isloading = false;
+  bool isLoading = false;
   final TextEditingController _emailController = TextEditingController();
 
   ApiService _apiService = ApiService();
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: isloading == true
-            ? LoadingIndicator()
-            : SingleChildScrollView(
-                child: Padding(
-                  padding: EdgeInsets.all(16.0),
-                  child: Form(
-                    key: _formKey,
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        SizedBox(
-                          height: 130.0,
-                        ),
-                        Text(
-                          'Forgot Password?',
-                          style: TextStyle(
-                            color: Colors.black,
-                            fontSize: 22.0,
-                            fontWeight: FontWeight.w600,
-                            fontFamily: 'Montserrat',
-                            letterSpacing: 0.03,
-                          ),
-                        ),
-                        SizedBox(
-                          height: 5,
-                        ),
-                        Text(
-                          'Enter your email address linked to your account. A password reset link will be sent to the email',
-                          style: TextStyle(
-                            color: Color.fromRGBO(0, 0, 0, 0.6),
-                            fontSize: 14,
-                            fontFamily: 'Montserrat',
-                          ),
-                        ),
-                        SizedBox(
-                          height: 22,
-                        ),
-                        Text(
-                          'Email Address',
-                          style: TextStyle(
-                            color: Color.fromRGBO(0, 0, 0, 0.6),
-                            fontSize: 13,
-                            fontWeight: FontWeight.w300,
-                            fontFamily: 'Montserrat',
-                          ),
-                        ),
-                        SizedBox(
-                          height: 5,
-                        ),
-                        TextFormField(
-                          controller: _emailController,
-                          validator: Validators.compose([
-                            Validators.required('Input Email Address'),
-                            Validators.email('Invalid Email Address'),
-                          ]),
-                          style: TextStyle(
-                            color: Color(0xFF2B2B2B),
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                            fontFamily: 'Montserrat',
-                          ),
-                          keyboardType: TextInputType.emailAddress,
-                          decoration: InputDecoration(
-                            contentPadding: EdgeInsets.all(15),
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(5),
-                              borderSide: BorderSide(
-                                color: Color(0xFFC8C8C8),
-                                width: 1,
-                              ),
-                            ),
-                            focusedBorder: OutlineInputBorder(),
-                            errorStyle: TextStyle(height: 0.5),
-                          ),
-                        ),
-                        SizedBox(
-                          height: 213,
-                        ),
-                        SizedBox(
-                          width: double.infinity,
-                          height: 45,
-                          child: FlatButton(
-                            onPressed: () async {
-                              if (_formKey.currentState.validate()) {
-                                await verifyUserToResetPassword();
-                              }
-                            },
-                            padding: EdgeInsets.all(10),
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(5),
-                            ),
-                            child: Text(
-                              'Request reset OTP',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontFamily: 'Montserrat',
-                                fontSize: 16,
-                                letterSpacing: 0.3,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                            color: Color(0xFF0B57A7),
-                          ),
-                        ),
-                        SizedBox(
-                          height: 30,
-                        ),
-                      ],
+        body: SingleChildScrollView(
+      child: Padding(
+        padding: EdgeInsets.all(16.0),
+        child: Form(
+          key: _formKey,
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              SizedBox(
+                height: 130.0,
+              ),
+              Text(
+                'Forgot Password?',
+                style: TextStyle(
+                  color: Colors.black,
+                  fontSize: 22.0,
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'Montserrat',
+                  letterSpacing: 0.03,
+                ),
+              ),
+              SizedBox(
+                height: 5,
+              ),
+              Text(
+                'Enter your email address linked to your account. A password reset link will be sent to the email',
+                style: TextStyle(
+                  color: Color.fromRGBO(0, 0, 0, 0.6),
+                  fontSize: 14,
+                  fontFamily: 'Montserrat',
+                ),
+              ),
+              SizedBox(
+                height: 22,
+              ),
+              Text(
+                'Email Address',
+                style: TextStyle(
+                  color: Color.fromRGBO(0, 0, 0, 0.6),
+                  fontSize: 13,
+                  fontWeight: FontWeight.w300,
+                  fontFamily: 'Montserrat',
+                ),
+              ),
+              SizedBox(
+                height: 5,
+              ),
+              TextFormField(
+                controller: _emailController,
+                validator: Validators.compose([
+                  Validators.required('Input Email Address'),
+                  Validators.email('Invalid Email Address'),
+                ]),
+                style: TextStyle(
+                  color: Color(0xFF2B2B2B),
+                  fontSize: 14,
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'Montserrat',
+                ),
+                keyboardType: TextInputType.emailAddress,
+                decoration: InputDecoration(
+                  contentPadding: EdgeInsets.all(15),
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(5),
+                    borderSide: BorderSide(
+                      color: Color(0xFFC8C8C8),
+                      width: 1,
                     ),
                   ),
+                  focusedBorder: OutlineInputBorder(),
+                  errorStyle: TextStyle(height: 0.5),
                 ),
-              ));
+              ),
+              SizedBox(
+                height: 213,
+              ),
+              SizedBox(
+                width: double.infinity,
+                height: 45,
+                child: FlatButton(
+                  onPressed: () async {
+                    if (_formKey.currentState.validate()) {
+                      await verifyUserToResetPassword();
+                    }
+                  },
+                  padding: EdgeInsets.all(10),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(5),
+                  ),
+                  child: isLoading
+                      ? ButtonLoadingIndicator(
+                          color: Colors.white,
+                          width: 20,
+                          height: 20,
+                        )
+                      : Text(
+                          'Request reset OTP',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontFamily: 'Montserrat',
+                            fontSize: 16,
+                            letterSpacing: 0.3,
+                            fontWeight: FontWeight.w600,
+                          ),
+                        ),
+                  color: Color(0xFF0B57A7),
+                ),
+              ),
+              SizedBox(
+                height: 30,
+              ),
+            ],
+          ),
+        ),
+      ),
+    ));
   }
 
   verifyUserToResetPassword() async {
     _formKey.currentState.save();
     setState(() {
-      isloading = true;
+      isLoading = true;
     });
     print('im res');
     Response response =
@@ -170,7 +175,7 @@ class _ForgotPasswordState extends State<ForgotPassword> {
 
       var error = res['error'];
       setState(() {
-        isloading = false;
+        isLoading = false;
       });
       Fluttertoast.showToast(
         msg: error,
@@ -178,9 +183,9 @@ class _ForgotPasswordState extends State<ForgotPassword> {
         toastLength: Toast.LENGTH_LONG,
         backgroundColor: Colors.red,
       );
-    }else{
+    } else {
       setState(() {
-        isloading = false;
+        isLoading = false;
       });
       Fluttertoast.showToast(
         msg: 'Sorry an error occured try again',
