+++ digital_receipt/lib/widgets/create_receipt_step0.dart
@@ -280,52 +280,6 @@ class _CreateReceiptStep0State extends State<CreateReceiptStep0> {
               SizedBox(
                 height: 10,
               ),
-/*
-              DropdownButtonFormField<Customer>(
-                value: selectedCustomer,
-                onChanged: (Customer value) {
-                  setState(() {
-                    selectedCustomer = value;
-                  });
-                },
-                validator: (value) {
-                  if (_nameController.text == null ||
-                      _nameController.text.isEmpty) {
-                    if (value == null) {
-                      return "Select a customer or enter a new one";
-                    }
-                  }
-                  return null;
-                },
-                items: customers.map((Customer customer) {
-                  return DropdownMenuItem<Customer>(
-                    value: customer.email != null ? customer : null,
-                    child: Text(
-                      customer.name,
-                    ),
-                  );
-                }).toList(),
-                iconEnabledColor: Color.fromRGBO(0, 0, 0, 0.87),
-                decoration: InputDecoration(
-                  contentPadding: EdgeInsets.all(15),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(5),
-                    borderSide: BorderSide(
-                      color: Color(0xFFC8C8C8),
-                      width: 1.5,
-                    ),
-                  ),
-                  focusedBorder: OutlineInputBorder(),
-                  //hintText: hintText,
-                  hintStyle: TextStyle(
-                    color: Color(0xFF979797),
-                    fontSize: 14,
-                    fontWeight: FontWeight.w500,
-                    fontFamily: 'Montserrat',
-                  ),
-                ),
-              ),
-*/
               GestureDetector(
                 onTap: () async {
                   showDialog(
@@ -549,223 +503,229 @@ class _CreateReceiptStep0State extends State<CreateReceiptStep0> {
               SizedBox(
                 height: 25,
               ),
-                DropdownButtonFormField(
-                items: Currency.currencyList().map<DropdownMenuItem<Currency>>((curr) => DropdownMenuItem(value: curr, child: Row(children: <Widget>[
-                  Text(curr.flag),
-                  SizedBox(width:7),
-                  Text(curr.currencyName),
-                  SizedBox(width:7),
-                  Text(curr.currencySymbol),
-                ],),
-                )).toList(),
-                onChanged: (Currency currency) {
-                  _changeCurrency(currency);
-                                  },
-                                  iconDisabledColor: Color.fromRGBO(0, 0, 0, 0.87),
-                                  decoration: InputDecoration(
-                                    contentPadding: EdgeInsets.all(15),
-                                    enabledBorder: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(5),
-                                      borderSide: BorderSide(
-                                        color: Color(0xFFC8C8C8),
-                                        width: 1.5,
-                                      ),
-                                    ),
-                                    focusedBorder: OutlineInputBorder(),
-                                    //hintText: hintText,
-                                    hintStyle: TextStyle(
-                                      color: Color(0xFF979797),
-                                      fontSize: 14,
-                                      fontWeight: FontWeight.w500,
-                                      fontFamily: 'Montserrat',
-                                    ),
-                                  ),
-                                  hint: Text(
-                                    'Select currency',
-                                    style: TextStyle(
-                                      fontFamily: 'Montserrat',
-                                      fontWeight: FontWeight.w500,
-                                      letterSpacing: 0.3,
-                                      fontSize: 14,
-                                      color: Color(0xFF1B1B1B),
-                                    ),
-                                  ),
-                                ), 
-                                SizedBox(
-                                  height: 45,
-                                ),
-                                SubmitButton(
-                                  title: 'Next',
-                                  textColor: Colors.white,
-                                  backgroundColor: Color(0xFF0B57A7),
-                                  onPressed: () {
-                                    if (_formKey.currentState.validate()) {
-                                      _formKey.currentState.save();
-                  
-                                      Provider.of<Receipt>(context, listen: false)
-                                          .setCategory(selectedCategory);
-                  
-                                      if (selectedCustomer == null) {
-                                        Provider.of<Receipt>(context, listen: false).setCustomer(
-                                          Customer(
-                                            name: _customerName,
-                                            email: _customerEmail,
-                                            phoneNumber: _customerPNumber,
-                                            address: _customerAddress,
-                                          ),
-                                        );
-                                      } else {
-                                        Provider.of<Receipt>(context, listen: false)
-                                            .setCustomer(selectedCustomer);
-                                      }
-                  
-                                      print(Provider.of<Receipt>(context, listen: false));
-                                      widget.carouselController.animateToPage(1);
-                                    }
-                                  },
-                                )
-                              ],
-                            ),
-                          ),
-                        ),
-                      );
-                    }
-                  
-                    @override
-                    void dispose() {
-                      _nameController.dispose();
-                      _emailController.dispose();
-                      _addressController.dispose();
-                      _pNumberController.dispose();
-                      super.dispose();
-                    }
-                  
-                    void _selectCustomerDropdown(BuildContext context,
-                        {Function(Customer) onSubmit}) {
-                      Dialog simpleDialog = Dialog(
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(5),
-                        ),
-                        child: Padding(
-                          padding: const EdgeInsets.all(16),
-                          child: Column(
+              DropdownButtonFormField(
+                items: Currency.currencyList()
+                    .map<DropdownMenuItem<Currency>>((curr) => DropdownMenuItem(
+                          value: curr,
+                          child: Row(
                             children: <Widget>[
-                              TextFormField(
-                                decoration: InputDecoration(
-                                  hintText: "Search customer",
-                                  hintStyle: TextStyle(
-                                      color: Color.fromRGBO(0, 0, 0, 0.38),
-                                      fontFamily: 'Montserrat'),
-                                  prefixIcon: IconButton(
-                                    icon: Icon(Icons.search),
-                                    color: Color.fromRGBO(0, 0, 0, 0.38),
-                                    onPressed: () {},
-                                  ),
-                                  border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: Color.fromRGBO(0, 0, 0, 0.12),
-                                      width: 1,
-                                    ),
-                                  ),
-                                  contentPadding: EdgeInsets.all(15),
-                                  enabledBorder: OutlineInputBorder(
-                                    borderRadius: BorderRadius.circular(5),
-                                    borderSide: BorderSide(
-                                      color: Color(0xFFC8C8C8),
-                                      width: 1.5,
-                                    ),
-                                  ),
-                                ),
-                              ),
-                              SizedBox(height: 20),
-                              Expanded(
-                                child: ListView.builder(
-                                  shrinkWrap: true,
-                                  itemCount: customers.length,
-                                  itemBuilder: (context, index) {
-                                    return InkWell(
-                                      onTap: () {
-                                        onSubmit(customers[index]);
-                                        Navigator.pop(context);
-                                      },
-                                      child: ContactCard(
-                                        receiptTitle: customers[index].name,
-                                        subtitle: customers[index].phoneNumber,
-                                      ),
-                                    );
-                                  },
-                                ),
-                              ),
+                              Text(curr.flag),
+                              SizedBox(width: 7),
+                              Text(curr.currencyName),
+                              SizedBox(width: 7),
+                              Text(curr.currencySymbol),
                             ],
                           ),
+                        ))
+                    .toList(),
+                onChanged: (Currency currency) {
+                  _changeCurrency(currency);
+                },
+                iconDisabledColor: Color.fromRGBO(0, 0, 0, 0.87),
+                decoration: InputDecoration(
+                  contentPadding: EdgeInsets.all(15),
+                  enabledBorder: OutlineInputBorder(
+                    borderRadius: BorderRadius.circular(5),
+                    borderSide: BorderSide(
+                      color: Color(0xFFC8C8C8),
+                      width: 1.5,
+                    ),
+                  ),
+                  focusedBorder: OutlineInputBorder(),
+                  //hintText: hintText,
+                  hintStyle: TextStyle(
+                    color: Color(0xFF979797),
+                    fontSize: 14,
+                    fontWeight: FontWeight.w500,
+                    fontFamily: 'Montserrat',
+                  ),
+                ),
+                hint: Text(
+                  'Select currency',
+                  style: TextStyle(
+                    fontFamily: 'Montserrat',
+                    fontWeight: FontWeight.w500,
+                    letterSpacing: 0.3,
+                    fontSize: 14,
+                    color: Color(0xFF1B1B1B),
+                  ),
+                ),
+              ),
+              SizedBox(
+                height: 45,
+              ),
+              SubmitButton(
+                title: 'Next',
+                textColor: Colors.white,
+                backgroundColor: Color(0xFF0B57A7),
+                onPressed: () {
+                  if (_formKey.currentState.validate()) {
+                    _formKey.currentState.save();
+
+                    Provider.of<Receipt>(context, listen: false)
+                        .setCategory(selectedCategory);
+
+                    if (selectedCustomer == null) {
+                      Provider.of<Receipt>(context, listen: false).setCustomer(
+                        Customer(
+                          name: _customerName,
+                          email: _customerEmail,
+                          phoneNumber: _customerPNumber,
+                          address: _customerAddress,
                         ),
                       );
-                      showDialog(
-                          context: context, builder: (BuildContext context) => simpleDialog);
+                    } else {
+                      Provider.of<Receipt>(context, listen: false)
+                          .setCustomer(selectedCustomer);
                     }
-                  
-                    Widget ContactCard({String receiptNo, total, date, receiptTitle, subtitle}) {
-                      return SizedBox(
-                        child: Column(
-                          children: <Widget>[
-                            Container(
-                              width: double.infinity,
-                              decoration: BoxDecoration(
-                                color: Color(0xFF539C30),
-                                borderRadius: BorderRadius.circular(10),
-                              ),
-                              child: Container(
-                                margin: EdgeInsets.only(left: 5.0),
-                                decoration: BoxDecoration(
-                                  color: Color(0xFFE2EAF3),
-                                  borderRadius: BorderRadius.circular(10),
-                                ),
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    Padding(
-                                      padding: EdgeInsets.fromLTRB(10.0, 5.0, 5.0, 5.0),
-                                      child: Text(
-                                        "$receiptTitle",
-                                        style: TextStyle(
-                                          color: Color.fromRGBO(0, 0, 0, 0.87),
-                                          fontSize: 14,
-                                          fontWeight: FontWeight.w600,
-                                          fontFamily: 'Montserrat',
-                                          letterSpacing: 0.03,
-                                        ),
-                                      ),
-                                    ),
-                                    Padding(
-                                      padding: EdgeInsets.fromLTRB(10.0, 5.0, 102.0, 5.0),
-                                      child: Text(
-                                        "$subtitle",
-                                        maxLines: 2,
-                                        overflow: TextOverflow.ellipsis,
-                                        style: TextStyle(
-                                          color: Colors.black,
-                                          fontSize: 14,
-                                          fontWeight: FontWeight.w100,
-                                          height: 1.43,
-                                          fontFamily: 'Montserrat',
-                                          letterSpacing: 0.3,
-                                        ),
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                            SizedBox(
-                              height: 20,
-                            ),
-                          ],
-                        ),
-                      );
-                    }
-                  
-                    void _changeCurrency(Currency currency) {
-                      print(currency.currencyName);
-                    }
+
+                    print(Provider.of<Receipt>(context, listen: false));
+                    widget.carouselController.animateToPage(1);
+                  }
+                },
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    _nameController.dispose();
+    _emailController.dispose();
+    _addressController.dispose();
+    _pNumberController.dispose();
+    super.dispose();
+  }
+
+  void _selectCustomerDropdown(BuildContext context,
+      {Function(Customer) onSubmit}) {
+    Dialog simpleDialog = Dialog(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(5),
+      ),
+      child: Padding(
+        padding: const EdgeInsets.all(16),
+        child: Column(
+          children: <Widget>[
+            TextFormField(
+              decoration: InputDecoration(
+                hintText: "Search customer",
+                hintStyle: TextStyle(
+                    color: Color.fromRGBO(0, 0, 0, 0.38),
+                    fontFamily: 'Montserrat'),
+                prefixIcon: IconButton(
+                  icon: Icon(Icons.search),
+                  color: Color.fromRGBO(0, 0, 0, 0.38),
+                  onPressed: () {},
+                ),
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(5),
+                  borderSide: BorderSide(
+                    color: Color.fromRGBO(0, 0, 0, 0.12),
+                    width: 1,
+                  ),
+                ),
+                contentPadding: EdgeInsets.all(15),
+                enabledBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(5),
+                  borderSide: BorderSide(
+                    color: Color(0xFFC8C8C8),
+                    width: 1.5,
+                  ),
+                ),
+              ),
+            ),
+            SizedBox(height: 20),
+            Expanded(
+              child: ListView.builder(
+                shrinkWrap: true,
+                itemCount: customers.length,
+                itemBuilder: (context, index) {
+                  return InkWell(
+                    onTap: () {
+                      onSubmit(customers[index]);
+                      Navigator.pop(context);
+                    },
+                    child: ContactCard(
+                      receiptTitle: customers[index].name,
+                      subtitle: customers[index].phoneNumber,
+                    ),
+                  );
+                },
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+    showDialog(
+        context: context, builder: (BuildContext context) => simpleDialog);
+  }
+
+  Widget ContactCard({String receiptNo, total, date, receiptTitle, subtitle}) {
+    return SizedBox(
+      child: Column(
+        children: <Widget>[
+          Container(
+            width: double.infinity,
+            decoration: BoxDecoration(
+              color: Color(0xFF539C30),
+              borderRadius: BorderRadius.circular(10),
+            ),
+            child: Container(
+              margin: EdgeInsets.only(left: 5.0),
+              decoration: BoxDecoration(
+                color: Color(0xFFE2EAF3),
+                borderRadius: BorderRadius.circular(10),
+              ),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Padding(
+                    padding: EdgeInsets.fromLTRB(10.0, 5.0, 5.0, 5.0),
+                    child: Text(
+                      "$receiptTitle",
+                      style: TextStyle(
+                        color: Color.fromRGBO(0, 0, 0, 0.87),
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                        fontFamily: 'Montserrat',
+                        letterSpacing: 0.03,
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding: EdgeInsets.fromLTRB(10.0, 5.0, 102.0, 5.0),
+                    child: Text(
+                      "$subtitle",
+                      maxLines: 2,
+                      overflow: TextOverflow.ellipsis,
+                      style: TextStyle(
+                        color: Colors.black,
+                        fontSize: 14,
+                        fontWeight: FontWeight.w100,
+                        height: 1.43,
+                        fontFamily: 'Montserrat',
+                        letterSpacing: 0.3,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+          SizedBox(
+            height: 20,
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _changeCurrency(Currency currency) {
+    print(currency.currencyName);
+  }
 }
